update login and signup pages
