+++ src/auth.js
@@ -1,0 +1,33 @@
+// import React from "react";
+
+
+
+
+
+
+
+
+
+//  const  handleSignup=(values)=>{ 
+//     const auth=localStorage.getItem('auth');
+//     console.log(auth);
+// if(auth===null){
+//     auth=[{'username':'aaa','password':'aaa'}];
+// }
+// if(values.password===values.password){
+//     const same=auth.filter(d=>d.values.username===username);
+//     if(same.length===0){
+//         auth =[...auth,{'userName':username,"password": password}];
+//         localStorage.setItem('auth',JSON.stringify(auth));
+//         setUsername('');
+//         setPassword('');
+//         setPassword2('');
+//     }else{
+//         alert('user exist!')
+//     }
+    
+// }else{
+//     alert('Password are not matching')
+// }
+//  }
+//  export default handleSignup;
