+++ src/index.js
@@ -1,11 +1,18 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
+import Login from "./Login";
+// import Signup from "./Signup"
 
 import App from "./App";
-import Login from "./Login";
+
+
+
+
+
+
 
 
 
 
 const root =ReactDOM.createRoot(document.getElementById('root'));
-root.render(<Login/>);
+root.render(<App/>);
