+++ src/vali-login-.js
@@ -2,18 +2,18 @@ import React from 'react';
 
 const validation=(values)=>{
     let errors={};
-    if(!values.userName){
-        errors.userName="User name is required."
+    if(!values.username){
+        errors.username="User name is required."
     }
         // if(!values.email){
         //     errors.email ="Email is required."
         // }else if(!/\S+@\S+\.\S+/.test(values.email)){
         //     errors.email="Email is invalid."
         // }
-        if(!values.password1){
-            errors.password1="Password is required.";
-        } else if(values.password1.length<5){
-                errors.password1='Password must be more then five characters'
+        if(!values.password){
+            errors.password="Password is required.";
+        } else if(values.password.length<5){
+                errors.password='Password must be more then five characters'
             }
             // if(!values.password2){
             //     errors.password2="Password is required.";
